fix: Fix the number type of the Package class weight and volume props

+++ 4P/paa/u2/transportadora/gabrielbatista_202300027249_transportadora.cpp
@@ -25,8 +25,8 @@ class Package {
 public:
   string code;
   float value = 0;
-  float weight = 0;
-  float volume = 0;
+  int weight = 0;
+  int volume = 0;
 };
 
 class PackageList {
